+++ apps/web/src/views/Migration/components/v3/Step2.tsx
@@ -68,7 +68,9 @@ export function Step2() {
 
   const activeV3Farms = farmsWithPrice.filter((farm) => farm.multiplier !== '0X')
 
-  const v2Pairs = useV2Pairs(liquidityTokensWithBalances.map(({ tokens }) => tokens))
+  const v2Pairs = useV2Pairs(
+    useMemo(() => liquidityTokensWithBalances.map(({ tokens }) => tokens), [liquidityTokensWithBalances]),
+  )
   const v2IsLoading =
     fetchingV2PairBalances ||
     v2Pairs?.length < liquidityTokensWithBalances.length ||