+++ packages/routing-sdk/src/index.test.ts
@@ -1,0 +1,13 @@
+import { expect, test } from 'vitest'
+import * as exportedNameSpaces from './index'
+
+test('exports', () => {
+  expect(Object.keys(exportedNameSpaces)).toMatchInlineSnapshot(`
+    [
+      "getNeighbour",
+      "getEdgeKey",
+      "createGraph",
+      "createPriceCalculator",
+    ]
+  `)
+})