+++ packages/farms/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @pancakeswap/farms
 
+## 1.0.64
+
+### Patch Changes
+
+- Updated dependencies [edf4640]
+  - @pancakeswap/chains@0.4.4
+  - @pancakeswap/price-api-sdk@1.2.11
+  - @pancakeswap/sdk@5.8.4
+  - @pancakeswap/tokens@0.6.18
+  - @pancakeswap/utils@6.0.6
+  - @pancakeswap/v3-sdk@3.8.5
+  - @pancakeswap/token-lists@0.0.11
+
+## 1.0.63
+
+### Patch Changes
+
+- 755bb18: New FURY Token. 2 BSC Farms. 3 zkSync Farms. New Gauges.
+- 502d961: Add MNT and mETH tokens with their farms and gauges on Ethereum
+- Updated dependencies [b57fd91]
+- Updated dependencies [e99c216]
+- Updated dependencies [755bb18]
+- Updated dependencies [502d961]
+  - @pancakeswap/tokens@0.6.17
+  - @pancakeswap/sdk@5.8.3
+  - @pancakeswap/price-api-sdk@1.2.10
+  - @pancakeswap/utils@6.0.5
+  - @pancakeswap/v3-sdk@3.8.4
+
 ## 1.0.62
 
 ### Patch Changes