+++ packages/games/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pancakeswap/games
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [edf4640]
+  - @pancakeswap/chains@0.4.4
+
 ## 0.0.6
 
 ### Patch Changes