+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pancakeswap/hooks
 
+## 0.0.28
+
+### Patch Changes
+
+- @pancakeswap/utils@6.0.6
+
 ## 0.0.27
 
 ### Patch Changes