+++ packages/multicall/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @pancakeswap/multicall
 
+## 3.5.4
+
+### Patch Changes
+
+- Updated dependencies [edf4640]
+  - @pancakeswap/chains@0.4.4
+  - @pancakeswap/sdk@5.8.4
+
+## 3.5.3
+
+### Patch Changes
+
+- Updated dependencies [e99c216]
+  - @pancakeswap/sdk@5.8.3
+
 ## 3.5.2
 
 ### Patch Changes