+++ packages/permit2-sdk/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @pancakeswap/permit2-sdk
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [edf4640]
+  - @pancakeswap/chains@0.4.4
+  - @pancakeswap/sdk@5.8.4
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [e99c216]
+  - @pancakeswap/sdk@5.8.3
+
 ## 1.0.4
 
 ### Patch Changes