+++ packages/pools/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @pancakeswap/pools
 
+## 5.1.24
+
+### Patch Changes
+
+- Updated dependencies [edf4640]
+  - @pancakeswap/chains@0.4.4
+  - @pancakeswap/tokens@0.6.18
+  - @pancakeswap/token-lists@0.0.11
+
+## 5.1.23
+
+### Patch Changes
+
+- Updated dependencies [b57fd91]
+- Updated dependencies [755bb18]
+- Updated dependencies [502d961]
+  - @pancakeswap/tokens@0.6.17
+
 ## 5.1.22
 
 ### Patch Changes