+++ packages/prediction/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @pancakeswap/prediction
 
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies [edf4640]
+  - @pancakeswap/chains@0.4.4
+  - @pancakeswap/pools@5.1.24
+  - @pancakeswap/sdk@5.8.4
+  - @pancakeswap/tokens@0.6.18
+
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [b57fd91]
+- Updated dependencies [e99c216]
+- Updated dependencies [755bb18]
+- Updated dependencies [502d961]
+  - @pancakeswap/tokens@0.6.17
+  - @pancakeswap/sdk@5.8.3
+  - @pancakeswap/pools@5.1.23
+
 ## 1.0.11
 
 ### Patch Changes