+++ packages/smart-router/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @pancakeswap/smart-router
 
+## 6.0.10
+
+### Patch Changes
+
+- Updated dependencies [edf4640]
+  - @pancakeswap/chains@0.4.4
+  - @pancakeswap/gauges@1.2.6
+  - @pancakeswap/multicall@3.5.4
+  - @pancakeswap/stable-swap-sdk@1.0.10
+  - @pancakeswap/sdk@5.8.4
+  - @pancakeswap/tokens@0.6.18
+  - @pancakeswap/v3-sdk@3.8.5
+  - @pancakeswap/token-lists@0.0.11
+
+## 6.0.9
+
+### Patch Changes
+
+- Updated dependencies [b57fd91]
+- Updated dependencies [e99c216]
+- Updated dependencies [755bb18]
+- Updated dependencies [502d961]
+  - @pancakeswap/tokens@0.6.17
+  - @pancakeswap/sdk@5.8.3
+  - @pancakeswap/gauges@1.2.5
+  - @pancakeswap/stable-swap-sdk@1.0.9
+  - @pancakeswap/multicall@3.5.3
+  - @pancakeswap/v3-sdk@3.8.4
+
 ## 6.0.8
 
 ### Patch Changes